+++ src/store/actions/actionTypes.js
@@ -3,4 +3,5 @@ const actions = {
     UPDATE_PRICE: 'UPDATE_PRICE'
 }
 
+
 export default actions;
