+++ src/store/utils.js
@@ -1,0 +1,6 @@
+export const updateObject = (oldObject, updatedValues) => {
+    return {
+        ...oldObject,
+        ...updatedValues
+    }
+}
